refactor(renderer): split texture management out of mod.rs
The unload_texture fix grew renderer/mod.rs past its ratcheted
baseline; the correct response is shrinking, not baseline-bumping.
register_texture* / update / unload / filter / evict_model_cache move
to renderer/texture_store.rs (341 lines); mod.rs ratchets down
13956 -> 13660.
