+++ tools/file-lines-baseline.json
@@ -1,5 +1,5 @@
 {
-  "native/shared/src/renderer/mod.rs": 13956,
+  "native/shared/src/renderer/mod.rs": 13660,
   "native/shared/src/renderer/shaders.rs": 5861,
   "native/web/src/lib.rs": 2146,
   "native/shared/src/renderer/material_system.rs": 2118,
